Reorganize, add apk, and update Readme
