--- 2Dgame/Assets/Prefabs/Platforms.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: e356bee0d638b344da8305d9efaef060
-folderAsset: yes
-timeCreated: 1493841202
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
