--- 2Dgame/Assets/Prefabs/Platforms/FlipPlatform.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 5ce1ea63bf9a69f49a7ef649d944e040
-timeCreated: 1494866550
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
