--- 2Dgame/Assets/Prefabs/Platforms/Ground.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0ea57182f091a4c6b8a2c027ab74702f
-timeCreated: 1494948408
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
