--- 2Dgame/Assets/Prefabs/Platforms/Moving Platform.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c446954ee41ede84cb6a1fd5830d1a9f
-timeCreated: 1493841190
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
