--- 2Dgame/Assets/Prefabs/Platforms/RegularPlatform.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 90a4d452dee8347d8ba47d46f1724782
-timeCreated: 1490883279
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
