+++ 2Dgame/Assets/Prefabs/Trampoline.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a70aee4fdbd20e843aefed4cbb7d9d8f
+timeCreated: 1493844810
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
