--- 2Dgame/Assets/Scripts/FlipPlatformScript.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: b0d760ff3e1050545a3ae4c7452ed05f
-timeCreated: 1494866048
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
