+++ 2Dgame/Assets/Scripts/Platform Scripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: cf111d2c06140430cba36fa722fae417
+folderAsset: yes
+timeCreated: 1494969580
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
