+++ 2Dgame/Assets/Scripts/Platform Scripts/PlatformMoverScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 08d6e4978f522c747bddc83f1871856c
+timeCreated: 1493840091
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
