+++ 2Dgame/Assets/Scripts/Platform Scripts/PlatformScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 6d7d09ce3b29f4853aeaa4b254d9b465
+timeCreated: 1493284233
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
