+++ 2Dgame/Assets/Scripts/Platform Scripts/TrampolineScript.cs
@@ -1,0 +1,7 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class TrampolineScript : PlatformScript {
+	public TrampolineScript() { jumpSpeed = 40f; }
+}
