+++ 2Dgame/Assets/Scripts/Scene Controllers/HighscoreScreenController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: f37af3eab4563432cb460ee4c59b5188
+timeCreated: 1493376214
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
