+++ 2Dgame/Assets/Scripts/Scene Controllers/SettingsScreenController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 81e28e43bd15c46618b6c9028325d708
+timeCreated: 1493376070
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
