--- 2Dgame/Assets/Scripts/SettingsScreenController.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 81e28e43bd15c46618b6c9028325d708
-timeCreated: 1493376070
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
