--- 2Dgame/Assets/Scripts/StartScreenController.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 5f3db99638750453fbc7940947f610bc
-timeCreated: 1493374779
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
