--- 2Dgame/Assets/Scripts/TrampolineScript.cs
@@ -1,7 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class TrampolineScript : PlatformScript {
-	public TrampolineScript() { jumpSpeed = 40f; }
-}
