--- 2Dgame/Assets/Scripts/TrampolineScript.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: b20fa00d2782f4e4b855d73e75b75b00
-timeCreated: 1494404941
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
